revert(rippling): remove the connectivity dataset; keep only the group catchment
The DfT connectivity-friction model didn't help in practice, so rip it out. Removed:
- friction/willingness reach model (friction.go), connectivity index (connectivity.go),
  Node.Conn, the uk_lsoa_connectivity.csv + build-connectivity pipeline (E&W/Scotland/NI),
  CONNECTIVITY_CSV wiring, friction params on /v1/isochrone, and the Connectivity-friction /
  Community-view toggles + friction overlay in the explorer.

Kept the one useful part — the **group catchment** tab, now plain travel-time only:
- multiSourceIsochrone (catchment.go): plain Dijkstra seeded from the group's whole boundary
  (via /v1/catchment?groupid=), so corridor reach into the group's edges still shows (the M62
  → Hull's western strip case). groupSeedNodes + /v1/groups/list unchanged.

Also fixed the 3-tab overflow: the mode buttons had default min-width:auto, so the long labels
forced each to content width and the 3rd tab spilled out of the pane. Added min-width:0 so
flex:1 shares the width and the labels wrap.

routing 60 pass, vitest 13710 pass.
