--- iznik-routing-go/data/build-connectivity/.gitignore
@@ -1,15 +1,0 @@
-# Transient build artifacts — regenerated by build.sh, never committed.
-# (The committed output is ../uk_lsoa_connectivity.csv, not anything in here.)
-node_modules/
-package-lock.json
-# package.json IS committed — it declares the NI build-time deps (proj4, xlsx) so a fresh
-# checkout can `npm install` without reading a script header.
-*.ods
-*.xlsx
-*.xls
-*.geojson
-*.zip
-lsoa_conn_codes.csv
-*_rows.csv
-/uk_lsoa_connectivity.csv
-connectivity_metrics*.ods
