+++ iznik-routing-go/main.go
@@ -36,17 +36,5 @@ func main() {
 	}
 	log.Printf("spatial-server: loaded %d nodes, %d edges", g.NodeCount(), len(g.Edges))
 
-	// DfT transport-connectivity, for the connectivity-friction reach model. Optional;
-	// unscored nodes (Conn==0, e.g. Scotland) fall back to the plain isochrone.
-	if path := getenv("CONNECTIVITY_CSV", ""); path != "" {
-		log.Printf("spatial-server: loading connectivity data from %s", path)
-		if ci := LoadConnectivity(path); ci == nil {
-			log.Printf("spatial-server: WARNING: failed to load connectivity data from %s", path)
-		} else {
-			TagConnectivity(g, ci)
-			log.Printf("spatial-server: connectivity data loaded and tagged onto nodes")
-		}
-	}
-
 	startServer(g)
 }
